Update result writer to fix the delete directory issue

+++ src/main/java/utils/ResultWriter.java
@@ -204,10 +204,12 @@ public class ResultWriter {
                 deleteDirectory(file);
             }
         }
-        boolean deleteResult = directoryToBeDeleted.delete();
-        if (!deleteResult) {
-            // throw new IOException("[ResultWriter.deleteDirectory] Failed to delete given directory: " + directoryToBeDeleted.getAbsolutePath());
-            // TODO: 
+        if (directoryToBeDeleted.exists()) {
+            boolean deleteResult = directoryToBeDeleted.delete();
+            if (!deleteResult) {
+                throw new IOException("[ResultWriter.deleteDirectory] Failed to delete given directory: " + directoryToBeDeleted.getAbsolutePath());
+            }
         }
+
     }
 }
